home and doc MVP page updated
