+++ src/client/components/LandingPage/carousel.tsx
@@ -1,120 +1,57 @@
-import React from 'react';
+import React, {useState} from 'react';
 import npmLogo from '../../assets/npm-logo.png';
 
-const Carousel: React.FC = () => {
+//Carousel type??
+// interface carouselTexts {
+//   string: string,
+//   string: string,
+//   string: string,
+// }
+
+// Array of carousel texts
+const Carousel = (): JSX.Element => {
+  const [currentIndex, setCurrentIndex] = useState(0);
+  const carouselTexts = [
+    'User Friendly UI with its own NPM package',
+    'Easy installation process',
+    'Simple Installation, recommended and utilized by Software Engineering teams',
+  ];
+
+  const handleNext = () => {
+    setCurrentIndex(prevIndex =>
+      prevIndex === carouselTexts.length - 1 ? 0 : prevIndex + 1
+    );
+  };
+
+  const handlePrev = () => {
+    setCurrentIndex(prevIndex =>
+      prevIndex === 0 ? carouselTexts.length - 1 : prevIndex - 1
+    );
+  };
+
   return (
-    <div id="carousel">
-      <div id="carousel-container">
-        <div id="carouselExampleIndicators" className="carousel slide">
-          <div className="carousel-indicators">
-            <button
-              type="button"
-              data-bs-target="#carouselExampleIndicators"
-              data-bs-slide-to="0"
-              className="active"
-              aria-current="true"
-              aria-label="Slide 1"
-            ></button>
-            <button
-              type="button"
-              data-bs-target="#carouselExampleIndicators"
-              data-bs-slide-to="1"
-              aria-label="Slide 2"
-            ></button>
-            <button
-              type="button"
-              data-bs-target="#carouselExampleIndicators"
-              data-bs-slide-to="2"
-              aria-label="Slide 3"
-            ></button>
-          </div>
-          <div className="carousel-inner">
-            <div className="carousel-item active">
-              <section className="cards-container">
-                <div className="card">
-                  <div className="card-body">
-                    <h5 className="card-title">Flake-guard feature 1</h5>
-                    <p className="card-text">Explanation & benefit.</p>
-                    <a href="#" className="btn btn-primary">
-                      Go to URL
-                    </a>
-                  </div>
-                </div>
-              </section>
-              <img
-                src=""
-                id="feature-1-logo"
-                // className="d-block w-100"
-                alt="..."
-              />
-            </div>
-            <div className="carousel-item">
-              <section className="cards-container">
-                <div className="card">
-                  <div className="card-body">
-                    <h5 className="card-title">Flake-guard feature 2 </h5>
-                    <p className="card-text">Explanation & benefit.</p>
-                    <a href="#" className="btn btn-primary">
-                      Go to URL
-                    </a>
-                  </div>
-                </div>
-              </section>
-              <img src="..." className="d-block w-100" alt="..." />
-            </div>
-            <div className="carousel-item">
-              <section className="cards-container">
-                <div className="card">
-                  <div className="card-body">
-                    <h5 className="card-title">NPM Package</h5>
-                    <p className="card-text">
-                      Easy to utilize npm package{' '}
-                    </p>{' '}
-                    <div className="registry-link">
-                      <a href="https://www.npmjs.com/package/flake-guard">
-                        <img src={npmLogo} alt="npm logo" id="social-logo" />
-                      </a>
-                    </div>
-                    <a
-                      href="https://www.npmjs.com/package/flake-guard"
-                      className="btn btn-primary"
-                    >
-                      NPM Registry
-                    </a>
-                  </div>
-                </div>
-              </section>
-              <img src="..." className="d-block w-100" alt="..." />
-            </div>
-          </div>
-          <button
-            className="carousel-control-prev"
-            type="button"
-            data-bs-target="#carouselExampleIndicators"
-            data-bs-slide="prev"
-          >
-            <span
-              className="carousel-control-prev-icon"
-              aria-hidden="true"
-            ></span>
-            <span className="visually-hidden">Previous</span>
+    <>
+      <div id="carousel-card" className="cards-container">
+        <div id="carousel-overlay">
+          {carouselTexts[currentIndex]}
+          {currentIndex === 1 && (
+            <img
+              src={npmLogo}
+              alt="npm logo"
+              style={{width: '70px', height: '70px'}}
+            />
+          )}
+        </div>
+        <div id="carousel-container">
+          <button id="carousel-prev" onClick={handlePrev}>
+            Prev
           </button>
-          <button
-            className="carousel-control-next"
-            type="button"
-            data-bs-target="#carouselExampleIndicators"
-            data-bs-slide="next"
-          >
-            <span
-              className="carousel-control-next-icon"
-              aria-hidden="true"
-            ></span>
-            <span className="visually-hidden">Next</span>
+          <button id="carousel-next" onClick={handleNext}>
+            Next
           </button>
         </div>
       </div>
-      <br />
-    </div>
+    </>
   );
 };
