+++ src/styles/styles.css
@@ -33,7 +33,7 @@ border-radius: 10px; /*  Rounded corners */
   color: white; /* Text color */
   /* font-size: 2rem; */
   font-family: Arial, sans-serif;
-  background: rgba(153, 150, 150, 0.5); /*  Semi-transparent background for better readability */
+  background: rgba(67, 67, 67, 0.5); /*  Semi-transparent background for better readability */
   padding: 20px;
   border-radius: 10px; /*  Rounded corners */
   }
@@ -53,7 +53,8 @@ border-radius: 10px; /*  Rounded corners */
   text-align: center;
   position: relative;
   width: 100%;
-  height: 100vh;
+  height: 250px;
+  zoom: unset;
 }
 .codeScreen-container {
   display: flex;
@@ -115,13 +116,9 @@ border-radius: 10px; /*  Rounded corners */
   border-radius: 25%;
 }
 #tagline-emphasize {
-  background: linear-gradient(
-    to right,
-    #00A86B,
-    #6610F2
-  ); /* Jade to Indigo gradient */
   background-clip: text; /* Standard */
-  color: transparent;
+  color: rgb(99, 192, 238);
+  font-size: 2rem;;
 }
 
 #social-logo {
@@ -136,9 +133,8 @@ border-radius: 10px; /*  Rounded corners */
   justify-content: center;
   text-align: center;
   color: white;
-  background-color: black;
   padding: 20px;
-  height: 20vh;
+  height: 30vh;
 }
 
 #carousel-container {
@@ -154,17 +150,23 @@ border-radius: 10px; /*  Rounded corners */
   background-color: white;
   color: black;
   border: none;
-  padding: 10px 20px;
+  /* padding: 10px 20px; */
+  margin: 10px;
   cursor: pointer;
+  width:150px;
+  height: 50px;
+  font-size: 30px;
 }
 
 #carousel-next {
-  text-align: end; /* This aligns text within the button to the end */
+  text-align: center; 
 }
 
 #carousel-overlay{
   color: white;
-  background-color: rgba(212, 54, 54, 0.639)
+  text-align: center;
+  justify-content: center;
+  align-items: center;
 }
 
 #feature-1-logo{
